requirements.txt
fastapi==0.95.2
langchain==0.0.173
pandas==2.0.1
python-dotenv==1.0.0
uvicorn==0.22.0
openai